Lazy Loading Tabs and Field Trigger Options
If you lazy load a tab which would only otherwise be displayed when all the fields within fulfil their required trigger options the tab is displayed anyway. The tab should only be displayed when the trigger options are complete, eg:

If you click on the **Employment Details** tab there are no fields until you flick the `is_employed` switch. Without lazy loading the tab wouldn't be displayed until the fields are triggered.